manifest: move Event into a file of its own
